+++ locales/en/notifications.json
@@ -73,7 +73,8 @@
   "ParsingCsv": "Parsing CSV file",
   "ParsingCsvWithLineNumber": "Parsing CSV file row",
   "StartingRoutingParallelServers": "Starting parallel routing servers",
-  "StoppingRoutingParallelServers": "Stoping parallel routing servers",
+  "StoppingRoutingParallelServers": "Stopping parallel routing servers",
+  "GeneratingBatchRoutingResults": "Generating batch routing result files",
   "GTFSExporterPreparation": "GTFS Export preparation",
   "GTFSExporterZipping": "GTFS Export compressing",
   "pending": "pending",
