+++ packages/transition-backend/src/services/transitRouting/types.ts
@@ -5,11 +5,19 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import { Feature, GeoJsonProperties, Geometry } from 'geojson';
-import { TransitRoutingResult } from 'transition-common/lib/services/transitRouting/TransitRoutingResult';
+import { ResultsByMode } from 'transition-common/lib/services/transitRouting/TransitRoutingCalculator';
+
+export type OdTripRouteResult = {
+    uuid: string;
+    internalId: string;
+    origin?: GeoJSON.Point;
+    destination?: GeoJSON.Point;
+    results?: ResultsByMode;
+    error?: string | { error: string; errorCode: string };
+};
 
 export type OdTripRouteOutput = {
     csv?: string[] | undefined;
     csvDetailed?: string[] | undefined;
     geometries?: Feature<Geometry, GeoJsonProperties>[] | undefined;
-    result?: TransitRoutingResult | undefined;
 };
